Hospital Management System/requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
python-dotenv==1.0.1
Werkzeug==3.0.3
reportlab==4.2.2
